server/unified/requirements.txt
Django==4.2.3
python-decouple==3.8
mysqlclient
djangorestframework~=3.14.0
django-cors-headers~=4.2.0